refactor: added utils directory, drafted README
