+++ src/datasetpreparator/sc2/sc2reset_replaypack_downloader/README.md
@@ -1,0 +1,38 @@
+# SC2ReSet Replaypack Downloader
+
+Utility script, downloads the contents of SC2ReSet replaypack from a Zenodo repository (https://doi.org/10.5281/zenodo.5575796).
+
+# CLI Usage
+
+Please keep in mind that the ```src/sc2reset_replaypack_downloader.py```  contains default flag values and can be customized with the following command line flags:
+```
+
+```
+
+## Citation
+
+SC2ReSet replaypack collection was formally introduced in:
+
+```bibtex
+@article{Białecki2023,
+  author  = {Bia{\l}ecki, Andrzej
+             and Jakubowska, Natalia
+             and Dobrowolski, Pawe{\l}
+             and Bia{\l}ecki, Piotr
+             and Krupi{\'{n}}ski, Leszek
+             and Szczap, Andrzej
+             and Bia{\l}ecki, Robert
+             and Gajewski, Jan},
+  title   = {SC2EGSet: StarCraft II Esport Replay and Game-state Dataset},
+  journal = {Scientific Data},
+  year    = {2023},
+  month   = {Sep},
+  day     = {08},
+  volume  = {10},
+  number  = {1},
+  pages   = {600},
+  issn    = {2052-4463},
+  doi     = {10.1038/s41597-023-02510-7},
+  url     = {https://doi.org/10.1038/s41597-023-02510-7}
+}
+```
